yatube/requirements.txt
Django==4.2
djangorestframework==3.14.0
django-filter==23.2
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.0
djoser==2.2.0
social-auth-app-django==5.4.0
